Fix for SDK-118
When extending FilterOutputStream, it's imperative to override all three write methods.  The array write versions of FilterOutputStream delegate to the single-byte version and therefore will cause huge performance issues if you don't delegate the array writes yourself.  Also, it's important to delegate to the wrapped ('out') stream and not super.